Create Simple_Trendilo_Strategy
This strategy is provided for demonstration purposes. It can be compared to the original indicator script by @dudeowns, which can be found here: https://www.tradingview.com/script/h5kMWewu-Trendilo-OPEN-SOURCE

The indicator was chosen largely at random for demonstration purposes; the same principals can be applied to any indicator or combination of indicators.